feat(*): Converted Heros to Service
Here’s what you achieved in this page:

You created a service class that can be shared by many components.
You used the ngOnInit lifecycle hook to get the hero data when the AppComponent activates.
You defined the HeroService as a provider for the AppComponent.
You designed the service to return a Future and the component to get the data from the Future.